requirements.txt
asgiref==3.11.1
attrs==25.4.0
coverage==7.13.4
Django==5.2
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.8.0
gunicorn==25.1.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
pytest-mock==3.15.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
ruff==0.15.5
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
